requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
requests==2.31.0
python-docx==1.1.0
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
textblob==0.17.1
vaderSentiment==3.3.2
scikit-learn==1.3.2
nltk==3.8.1
wordcloud==1.9.3

